[docs] re-sphinxing
