+++ .settings/org.eclipse.core.resources.prefs
@@ -1,5 +1,4 @@
 eclipse.preferences.version=1
-encoding//doc/old.conf.py=utf-8
 encoding//doc/source/conf.py=utf-8
 encoding//py_trees_ros/actions.py=utf-8
 encoding//py_trees_ros/battery.py=utf-8
@@ -13,14 +12,6 @@ encoding//py_trees_ros/programs/latched_echo.py=utf-8
 encoding//py_trees_ros/programs/tree_watcher.py=utf-8
 encoding//py_trees_ros/subscribers.py=utf-8
 encoding//py_trees_ros/trees.py=utf-8
-encoding//py_trees_ros/tutorials/behaviours.py=utf-8
-encoding//py_trees_ros/tutorials/eight.py=utf-8
-encoding//py_trees_ros/tutorials/five.py=utf-8
-encoding//py_trees_ros/tutorials/jobs.py=utf-8
-encoding//py_trees_ros/tutorials/qt.py=utf-8
-encoding//py_trees_ros/tutorials/seven.py=utf-8
-encoding//py_trees_ros/tutorials/six.py=utf-8
-encoding//py_trees_ros/tutorials/two.py=utf-8
 encoding//py_trees_ros/utilities.py=utf-8
 encoding//py_trees_ros/visitors.py=utf-8
 encoding//tests/test_battery.py=utf-8
