+++ doc/source/conf.py
@@ -79,7 +79,7 @@ napoleon_use_ivar = True
 
 # Example configuration for intersphinx: refer to the Python standard library.
 intersphinx_mapping = {
-    'https://docs.python.org/': None
+    'https://docs.python.org/': ('https://docs.python.org/3', None)
 }
 
 html_theme = 'sphinx_rtd_theme'
