+++ doc/source/features.rst
@@ -1,10 +1,6 @@
 Features
 ========
 
-.. seealso::
-    The :ref:`Tutorials <py-trees-ros-tutorials-section>` where most of the following
-    is demonstrated.
-
 Behaviours
 ----------
 
@@ -39,7 +35,3 @@ You can visualise the trees in ROS via either the ascii tree publishers:
    rostopic echo /tree/ascii/snapshot
 
 .. image:: images/ascii-snapshot.png
-
-or the `rqt_py_trees` plugin which tunes in to the `~log/tree` topic.
-
-.. image:: images/rqt-py-trees.png
