+++ graveyard/doc/modules.rst
@@ -1,0 +1,152 @@
+.. _modules-section-label:
+
+Module API
+==========
+
+py_trees_ros
+------------
+
+.. automodule:: py_trees_ros
+   :synopsis: is your robot misbehaving?
+
+py_trees_ros.actions
+--------------------
+
+.. automodule:: py_trees_ros.actions
+    :members:
+    :show-inheritance:
+    :synopsis: action server/client behaviours
+
+py_trees_ros.battery
+--------------------
+
+.. automodule:: py_trees_ros.battery
+    :members:
+    :show-inheritance:
+    :synopsis: helpers for getting the most out of your battery
+
+py_trees_ros.blackboard
+-----------------------
+
+.. automodule:: py_trees_ros.blackboard
+    :members:
+    :show-inheritance:
+    :synopsis: the ros blackboard and the blackboard watcher
+
+py_trees_ros.conversions
+------------------------
+
+.. automodule:: py_trees_ros.conversions
+    :members:
+    :show-inheritance:
+    :synopsis: py_trees - ros messages conversion library
+
+py_trees_ros.subscribers
+------------------------
+
+.. automodule:: py_trees_ros.subscribers
+    :members:
+    :show-inheritance:
+    :synopsis: put a lid on the asynchronicity
+
+py_trees_ros.trees
+------------------
+
+.. automodule:: py_trees_ros.trees
+    :members:
+    :show-inheritance:
+    :synopsis: extensions with ros handles for debugging, visualisation messages and more
+
+py_trees.utilities
+------------------
+
+.. automodule:: py_trees_ros.utilities
+    :members:
+    :show-inheritance:
+    :synopsis: assorted utility functions
+
+py_trees.visitors
+-----------------
+
+.. automodule:: py_trees_ros.visitors
+    :members:
+    :show-inheritance:
+    :synopsis: entities that visit behaviours as a tree is traversed
+
+py_trees_ros.mock
+-----------------
+
+.. automodule:: py_trees_ros.mock
+   :synopsis: utilities and components for mocking a robot
+
+py_trees_ros.mock.action_server
+-------------------------------
+
+.. automodule:: py_trees_ros.mock.action_server
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: reusable action server for mock components
+
+py_trees_ros.mock.battery
+-------------------------
+
+.. automodule:: py_trees_ros.mock.battery
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: mock the state of a battery component
+
+py_trees_ros.mock.move_base
+---------------------------
+
+.. automodule:: py_trees_ros.mock.move_base
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: mock the ROS navistack move base
+
+py_trees_ros.mock.rotate
+------------------------
+
+.. automodule:: py_trees_ros.mock.rotate
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: mock a very simple rotation action server
+
+py_trees_ros.tutorials
+----------------------
+
+.. automodule:: py_trees_ros.tutorials
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: everything to support the tutorials
+
+py_trees_ros.tutorials.behaviours
+---------------------------------
+
+.. automodule:: py_trees_ros.tutorials.behaviours
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: behaviours crafted specifically for the tutorials
+
+py_trees_ros.tutorials.jobs
+---------------------------
+
+.. automodule:: py_trees_ros.tutorials.jobs
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: behaviours crafted specifically for the tutorials
+
+py_trees_ros.tutorials.qt
+-------------------------
+
+.. automodule:: py_trees_ros.tutorials.qt
+    :members:
+    :special-members:
+    :show-inheritance:
+    :synopsis: graphical widgets specifically for interaction with the tutorials
