+++ py_trees_ros/__init__.py
@@ -14,18 +14,16 @@ ROS extensions, behaviours and utilities for py_trees.
 # Imports
 ##############################################################################
 
+from . import actions
 from . import battery
 from . import blackboard
-# from . import exceptions
+from . import conversions
+from . import exceptions
 from . import programs
 from . import subscribers
 from . import trees
-# from . import tutorials
 from . import utilities
-
-# from . import actions
-# from . import conversions
-# from . import visitors
+from . import visitors
 
 ##############################################################################
 # Version
